Add condition that if task doesn't exists throws error

+++ src/services/taskService/taskService.ts
@@ -29,6 +29,9 @@ export const updateTasks = async ( id: string, updatedTask: Task ): Promise<Task
 
 export const deleteTasks = async (id: string): Promise<void> => {
   const deleteTask = taskCollection.doc(id);
-  await deleteTask.get();
+  const doc = await deleteTask.get();
+  if (!doc.exists) {
+    throw new Error("Task not found");
+  }
   await deleteTask.delete();
 };
